Extract ElevenLabsTTSBase for the WebSocket TTS services
The text-to-speech and Text-to-Dialogue services share a WebSocket
connection and task lifecycle, settings-change dispatch, a word-timestamp
clock, and the shape of a synthesis request. `ElevenLabsTTSBase` in
`tts_base.py` owns those; each service supplies its endpoint URL and its
wire protocol through `_build_websocket_url`, `_send_context_init`,
`_send_text`, `_send_keepalive` and `_close_context`.

Settings dispatch runs off two declarative sets, `URL_FIELDS` and
`VOICE_SETTINGS_FIELDS`, which each service names on its own settings class.
`ElevenLabsTTSSettingsBase` declares them, so the contract is stated where
`_update_settings` reads it.

The alignment helpers move alongside the base class, so the Text-to-Dialogue
service no longer reaches into the text-to-speech module for them. They stay
importable from `pipecat.services.elevenlabs.tts`.
