+++ src/pipecat/services/elevenlabs/tts.py
@@ -10,7 +10,6 @@ This module provides WebSocket and HTTP-based TTS services using ElevenLabs API
 with support for streaming audio, word timestamps, and voice customization.
 """
 
-import asyncio
 import base64
 import json
 import warnings
@@ -24,7 +23,6 @@ from typing import (
 )
 
 import aiohttp
-import websockets
 from loguru import logger
 from pydantic import BaseModel
 from websockets.protocol import State
@@ -35,20 +33,43 @@ from pipecat.frames.frames import (
     InterruptionFrame,
     LLMFullResponseEndFrame,
     TTSAudioRawFrame,
-    TTSStartedFrame,
     TTSStoppedFrame,
 )
 from pipecat.processors.frame_processor import FrameDirection, FrameProcessorSetup
+from pipecat.services.elevenlabs.tts_base import (
+    ElevenLabsTTSBase,
+    ElevenLabsTTSSettingsBase,
+    _select_alignment,
+    _strip_utterance_leading_spaces,
+    _word_timestamps_include_inter_frame_spaces,
+    calculate_word_times,
+    language_to_elevenlabs_language,
+    output_format_from_sample_rate,
+)
 from pipecat.services.settings import TTSSettings
 from pipecat.services.tts_service import (
     TextAggregationMode,
     TTSService,
-    WebsocketTTSService,
 )
-from pipecat.transcriptions.language import Language, resolve_language
+from pipecat.transcriptions.language import Language
 from pipecat.utils.deprecation import deprecated
 from pipecat.utils.tracing.service_decorators import traced_tts
 from pipecat.utils.types import NOT_GIVEN, NotGiven, assert_given
+
+# Re-exported so the documented import path for these helpers keeps working.
+__all__ = [
+    "ELEVENLABS_CONTEXT_UNSUPPORTED_MODELS",
+    "ELEVENLABS_MULTILINGUAL_MODELS",
+    "ElevenLabsHttpTTSService",
+    "ElevenLabsHttpTTSSettings",
+    "ElevenLabsTTSService",
+    "ElevenLabsTTSSettings",
+    "PronunciationDictionaryLocator",
+    "build_elevenlabs_voice_settings",
+    "calculate_word_times",
+    "language_to_elevenlabs_language",
+    "output_format_from_sample_rate",
+]
 
 # Models that support language codes
 # The following models are excluded as they don't support language codes:
@@ -64,97 +85,6 @@ ELEVENLABS_MULTILINGUAL_MODELS = {
 ELEVENLABS_CONTEXT_UNSUPPORTED_MODELS = {
     "eleven_v3",
 }
-
-
-def language_to_elevenlabs_language(language: Language) -> str:
-    """Convert a Language enum to ElevenLabs language code.
-
-    Args:
-        language: The Language enum value to convert.
-
-    Returns:
-        The corresponding service language code. If ``language`` is not in
-        the verified mapping, falls back to the base language code (e.g.,
-        ``en`` from ``en-US``) and logs a warning (via
-        ``resolve_language(..., use_base_code=True)``).
-    """
-    LANGUAGE_MAP = {
-        Language.AR: "ar",
-        Language.BG: "bg",
-        Language.CS: "cs",
-        Language.DA: "da",
-        Language.DE: "de",
-        Language.EL: "el",
-        Language.EN: "en",
-        Language.ES: "es",
-        Language.FI: "fi",
-        Language.FIL: "fil",
-        Language.FR: "fr",
-        Language.HI: "hi",
-        Language.HR: "hr",
-        Language.HU: "hu",
-        Language.ID: "id",
-        Language.IT: "it",
-        Language.JA: "ja",
-        Language.KO: "ko",
-        Language.MS: "ms",
-        Language.NL: "nl",
-        Language.NO: "no",
-        Language.PL: "pl",
-        Language.PT: "pt",
-        Language.RO: "ro",
-        Language.RU: "ru",
-        Language.SK: "sk",
-        Language.SV: "sv",
-        Language.TA: "ta",
-        Language.TR: "tr",
-        Language.UK: "uk",
-        Language.VI: "vi",
-        Language.ZH: "zh",
-    }
-
-    return resolve_language(language, LANGUAGE_MAP, use_base_code=True)
-
-
-def output_format_from_sample_rate(sample_rate: int) -> str:
-    """Get the appropriate output format string for a given sample rate.
-
-    Args:
-        sample_rate: The audio sample rate in Hz.
-
-    Returns:
-        The ElevenLabs output format string.
-    """
-    match sample_rate:
-        case 8000:
-            return "pcm_8000"
-        case 16000:
-            return "pcm_16000"
-        case 22050:
-            return "pcm_22050"
-        case 24000:
-            return "pcm_24000"
-        case 32000:
-            return "pcm_32000"
-        case 44100:
-            return "pcm_44100"
-        case 48000:
-            return "pcm_48000"
-    logger.warning(
-        f"ElevenLabsTTSService: No output format available for {sample_rate} sample rate"
-    )
-    return "pcm_24000"
-
-
-def _is_chinese_or_japanese_language(language: str) -> bool:
-    """Check if the given language is Chinese or Japanese."""
-    base_lang = language.split("-")[0].lower()
-    return base_lang in {"zh", "ja"}
-
-
-def _word_timestamps_include_inter_frame_spaces(language: str | None) -> bool:
-    """Whether timestamp text should be treated as carrying its own spacing."""
-    return bool(language and _is_chinese_or_japanese_language(language))
 
 
 def build_elevenlabs_voice_settings(
@@ -207,7 +137,7 @@ class PronunciationDictionaryLocator(BaseModel):
 
 
 @dataclass
-class ElevenLabsTTSSettings(TTSSettings):
+class ElevenLabsTTSSettings(ElevenLabsTTSSettingsBase):
     """Settings for ElevenLabsTTSService.
 
     Fields that appear in the WebSocket URL (``voice``, ``model``,
@@ -269,140 +199,7 @@ class ElevenLabsHttpTTSSettings(TTSSettings):
     )
 
 
-def _select_alignment(
-    msg: Mapping[str, Any],
-    *,
-    normalized_key: str,
-    alignment_key: str,
-    prefer_normalized: bool,
-) -> Mapping[str, Any] | None:
-    """Pick the alignment field to use from a TTS message, with fallback.
-
-    ElevenLabs returns two alignment fields per chunk:
-
-    - ``normalized_key`` (``normalizedAlignment`` for WebSocket,
-      ``normalized_alignment`` for HTTP): the post-normalized form of what was
-      spoken - pronunciation-dictionary substitutions, text normalization, or
-      romanization of non-Latin scripts (e.g., Chinese rendered as pinyin).
-    - ``alignment_key`` (``alignment``): the original input characters.
-
-    Prefer ``normalized`` only when a pronunciation dictionary is configured -
-    that's the case where ``alignment`` has overlapping restarts that produce
-    duplicated/garbled words (issue #4316). Otherwise prefer ``alignment`` so
-    the LLM context preserves the original input rather than the normalized
-    form. Fall back to the other field if the preferred one is missing or
-    null - the API schema marks both as nullable.
-
-    Args:
-        msg: TTS response message from ElevenLabs.
-        normalized_key: Key for the normalized-alignment field on this transport.
-        alignment_key: Key for the original-alignment field on this transport.
-        prefer_normalized: True iff the caller is using pronunciation dictionaries.
-
-    Returns:
-        The chosen alignment dict, or ``None`` if both fields are absent/null.
-    """
-    if prefer_normalized:
-        return msg.get(normalized_key) or msg.get(alignment_key)
-    return msg.get(alignment_key) or msg.get(normalized_key)
-
-
-def _strip_utterance_leading_spaces(
-    alignment: Mapping[str, Any], keys: tuple[str, str, str], should_strip: bool
-) -> Mapping[str, Any]:
-    """Return alignment with utterance-leading space chars removed, if requested.
-
-    ElevenLabs Flash normalized alignment chunks can begin with a leading space
-    at the start of an utterance. Strip only utterance-leading spaces so bot
-    turn text does not start with whitespace. On subsequent chunks, however, a
-    leading space can be a real inter-word separator (Flash models commonly
-    split sentences this way), so it must be preserved for
-    ``calculate_word_times`` to flush any partial word carried over from the
-    previous chunk.
-
-    Args:
-        alignment: Alignment dict from the API.
-        keys: Tuple of (chars_key, start_times_key, durations_or_end_times_key)
-            naming the three parallel arrays - these differ between the
-            WebSocket and HTTP response schemas.
-        should_strip: Whether this is still utterance-leading alignment data.
-    """
-    chars_key, starts_key, tail_key = keys
-    chars = alignment.get(chars_key) or []
-    if should_strip and chars and chars[0] == " ":
-        strip_count = 0
-        while strip_count < len(chars) and chars[strip_count] == " ":
-            strip_count += 1
-
-        stripped = dict(alignment)
-        stripped[chars_key] = chars[strip_count:]
-        stripped[starts_key] = alignment.get(starts_key, [])[strip_count:]
-        stripped[tail_key] = alignment.get(tail_key, [])[strip_count:]
-        return stripped
-    return alignment
-
-
-def calculate_word_times(
-    alignment_info: Mapping[str, Any],
-    cumulative_time: float,
-    partial_word: str = "",
-    partial_word_start_time: float = 0.0,
-) -> tuple[list[tuple[str, float]], str, float]:
-    """Calculate word timestamps from character alignment information.
-
-    Args:
-        alignment_info: Character alignment data from ElevenLabs API.
-        cumulative_time: Base time offset for this chunk.
-        partial_word: Partial word carried over from previous chunk.
-        partial_word_start_time: Start time of the partial word.
-
-    Returns:
-        Tuple of (word_times, new_partial_word, new_partial_word_start_time):
-        - word_times: List of (word, timestamp) tuples for complete words
-        - new_partial_word: Incomplete word at end of chunk (empty if chunk ends with space)
-        - new_partial_word_start_time: Start time of the incomplete word
-    """
-    chars = alignment_info["chars"]
-    char_start_times_ms = alignment_info["charStartTimesMs"]
-
-    if len(chars) != len(char_start_times_ms):
-        logger.error(
-            f"calculate_word_times: length mismatch - chars={len(chars)}, times={len(char_start_times_ms)}"
-        )
-        return ([], partial_word, partial_word_start_time)
-
-    # Build words and track their start positions
-    words = []
-    word_start_times = []
-    current_word = partial_word  # Start with any partial word from previous chunk
-    word_start_time = partial_word_start_time if partial_word else None
-
-    for i, char in enumerate(chars):
-        if char == " ":
-            # End of current word
-            if current_word:  # Only add non-empty words
-                words.append(current_word)
-                word_start_times.append(word_start_time)
-                current_word = ""
-                word_start_time = None
-        else:
-            # Building a word
-            if word_start_time is None:  # First character of new word
-                # Convert from milliseconds to seconds and add cumulative offset
-                word_start_time = cumulative_time + (char_start_times_ms[i] / 1000.0)
-            current_word += char
-
-    # Build result for complete words
-    word_times = list(zip(words, word_start_times))
-
-    # Return any incomplete word at the end of this chunk
-    new_partial_word = current_word if current_word else ""
-    new_partial_word_start_time = word_start_time if word_start_time is not None else 0.0
-
-    return (word_times, new_partial_word, new_partial_word_start_time)
-
-
-class ElevenLabsTTSService(WebsocketTTSService):
+class ElevenLabsTTSService(ElevenLabsTTSBase):
     """ElevenLabs WebSocket-based TTS service with word timestamps.
 
     Provides real-time text-to-speech using ElevenLabs' WebSocket streaming API.
@@ -597,6 +394,9 @@ class ElevenLabsTTSService(WebsocketTTSService):
             default_settings.apply_update(settings)
 
         super().__init__(
+            api_key=api_key,
+            url=url,
+            enable_logging=enable_logging,
             text_aggregation_mode=text_aggregation_mode,
             aggregate_sentences=aggregate_sentences,
             push_text_frames=False,
@@ -606,9 +406,6 @@ class ElevenLabsTTSService(WebsocketTTSService):
             settings=default_settings,
             **kwargs,
         )
-
-        self._api_key = api_key
-        self._url = url
 
         # Init-only WebSocket URL params (not runtime-updatable).
         #
@@ -624,10 +421,7 @@ class ElevenLabsTTSService(WebsocketTTSService):
 
         self._auto_mode = auto_mode
         self._enable_ssml_parsing = enable_ssml_parsing
-        self._enable_logging = enable_logging
 
-        self._output_format = ""  # initialized in start()
-        self._voice_settings = self._set_voice_settings()
         if _pronunciation_dictionary_locators is not None:
             warnings.warn(
                 "`pronunciation_dictionary_locators` is deprecated since 1.6.0 and will be "
@@ -641,101 +435,14 @@ class ElevenLabsTTSService(WebsocketTTSService):
             )
         self._pronunciation_dictionary_locators = _pronunciation_dictionary_locators
 
-        self._cumulative_time = 0
-        # Track partial words that span across alignment chunks
-        self._partial_word = ""
-        self._partial_word_start_time = 0.0
         self._alignment_started_context_ids: set[str | None] = set()
 
         # Context IDs whose context-init has been sent, so the keepalive knows
         # which contexts are safe to target.
         self._context_init_sent: set[str] = set()
 
-        # Context management for v1 multi API
-        self._receive_task = None
-        self._keepalive_task = None
-
-    def can_generate_metrics(self) -> bool:
-        """Check if this service can generate processing metrics.
-
-        Returns:
-            True, as ElevenLabs service supports metrics generation.
-        """
-        return True
-
-    def language_to_service_language(self, language: Language) -> str | None:
-        """Convert a Language enum to ElevenLabs language format.
-
-        Args:
-            language: The language to convert.
-
-        Returns:
-            The ElevenLabs-specific language code, or None if not supported.
-        """
-        return language_to_elevenlabs_language(language)
-
     def _set_voice_settings(self):
         return build_elevenlabs_voice_settings(self._settings)
-
-    async def _update_settings(self, delta: TTSSettings) -> dict[str, Any]:
-        """Apply a settings delta, reconnecting as needed.
-
-        Uses the declarative ``URL_FIELDS`` and ``VOICE_SETTINGS_FIELDS``
-        sets on :class:`ElevenLabsTTSService.Settings` to decide whether to
-        reconnect the WebSocket or close the current audio context.
-
-        Args:
-            delta: A :class:`TTSSettings` (or ``ElevenLabsTTSService.Settings``) delta.
-
-        Returns:
-            Dict mapping changed field names to their previous values.
-        """
-        changed = await super()._update_settings(delta)
-
-        if not changed:
-            return changed
-
-        # Rebuild voice settings for next context
-        self._voice_settings = self._set_voice_settings()
-
-        url_changed = bool(changed.keys() & self.Settings.URL_FIELDS)
-        voice_settings_changed = bool(changed.keys() & self.Settings.VOICE_SETTINGS_FIELDS)
-
-        if url_changed:
-            logger.debug(
-                f"URL-level setting changed ({changed.keys() & self.Settings.URL_FIELDS}), "
-                f"reconnecting WebSocket"
-            )
-            await self._disconnect()
-            await self._connect()
-        elif voice_settings_changed:
-            logger.debug(
-                f"Voice settings changed ({changed.keys() & self.Settings.VOICE_SETTINGS_FIELDS}), "
-                f"closing current context to apply changes"
-            )
-            audio_contexts = self.get_audio_contexts()
-            if audio_contexts:
-                for ctx_id in audio_contexts:
-                    await self._close_context(ctx_id)
-                    self._reset_alignment_state(ctx_id)
-
-        if not url_changed:
-            # Reconnect applies all settings; only warn about fields not handled
-            # by voice settings or URL changes.
-            handled = self.Settings.URL_FIELDS | self.Settings.VOICE_SETTINGS_FIELDS
-            self._warn_unhandled_updated_settings(changed.keys() - handled)
-
-        return changed
-
-    async def setup(self, setup: FrameProcessorSetup):
-        """Set up the service and connect.
-
-        Args:
-            setup: Configuration object containing setup parameters.
-        """
-        await super().setup(setup)
-        self._output_format = output_format_from_sample_rate(self.sample_rate)
-        await self._connect()
 
     async def flush_audio(self, context_id: str | None = None):
         """Flush any pending audio and finalize the current context.
@@ -751,111 +458,35 @@ class ElevenLabsTTSService(WebsocketTTSService):
         msg = {"context_id": flush_id, "flush": True}
         await self._websocket.send(json.dumps(msg))
 
-    async def _connect(self):
-        await super()._connect()
+    def _build_websocket_url(self) -> str:
+        voice_id = self._settings.voice
+        model = self._settings.model
+        output_format = self._output_format
+        url = f"{self._url}/v1/text-to-speech/{voice_id}/multi-stream-input?model_id={model}&output_format={output_format}&auto_mode={str(self._auto_mode).lower()}"
 
-        await self._connect_websocket()
+        if self._enable_ssml_parsing is not None:
+            url += f"&enable_ssml_parsing={str(self._enable_ssml_parsing).lower()}"
 
-        if self._websocket and not self._receive_task:
-            self._receive_task = self.create_task(self._receive_task_handler(self._report_error))
+        if self._enable_logging is not None:
+            url += f"&enable_logging={str(self._enable_logging).lower()}"
 
-        if self._websocket and not self._keepalive_task:
-            self._keepalive_task = self.create_task(self._keepalive_task_handler())
+        if self._settings.apply_text_normalization is not None:
+            url += f"&apply_text_normalization={self._settings.apply_text_normalization}"
 
-    async def _disconnect(self):
-        await super()._disconnect()
-
-        if self._receive_task:
-            await self.cancel_task(self._receive_task)
-            self._receive_task = None
-
-        if self._keepalive_task:
-            await self.cancel_task(self._keepalive_task)
-            self._keepalive_task = None
-
-        await self._disconnect_websocket()
-
-    async def _connect_websocket(self):
-        try:
-            if self._websocket and self._websocket.state is State.OPEN:
-                return
-
-            logger.debug("Connecting to ElevenLabs")
-
-            voice_id = self._settings.voice
-            model = self._settings.model
-            output_format = self._output_format
-            url = f"{self._url}/v1/text-to-speech/{voice_id}/multi-stream-input?model_id={model}&output_format={output_format}&auto_mode={str(self._auto_mode).lower()}"
-
-            if self._enable_ssml_parsing is not None:
-                url += f"&enable_ssml_parsing={str(self._enable_ssml_parsing).lower()}"
-
-            if self._enable_logging is not None:
-                url += f"&enable_logging={str(self._enable_logging).lower()}"
-
-            if self._settings.apply_text_normalization is not None:
-                url += f"&apply_text_normalization={self._settings.apply_text_normalization}"
-
-            # Language can only be used with the ELEVENLABS_MULTILINGUAL_MODELS
-            language = self._settings.language
-            if model in ELEVENLABS_MULTILINGUAL_MODELS and language is not None:
-                url += f"&language_code={language}"
-                logger.debug(f"Using language code: {language}")
-            elif language is not None:
-                logger.warning(
-                    f"Language code [{language}] not applied. Language codes can only be used with multilingual models: {', '.join(sorted(ELEVENLABS_MULTILINGUAL_MODELS))}"
-                )
-
-            # Set max websocket message size to 16MB for large audio responses
-            self._websocket = await self._websocket_connect(
-                url, max_size=16 * 1024 * 1024, additional_headers={"xi-api-key": self._api_key}
+        # Language can only be used with the ELEVENLABS_MULTILINGUAL_MODELS
+        language = self._settings.language
+        if model in ELEVENLABS_MULTILINGUAL_MODELS and language is not None:
+            url += f"&language_code={language}"
+            logger.debug(f"Using language code: {language}")
+        elif language is not None:
+            logger.warning(
+                f"Language code [{language}] not applied. Language codes can only be used with multilingual models: {', '.join(sorted(ELEVENLABS_MULTILINGUAL_MODELS))}"
             )
 
-            await self._call_event_handler("on_connected")
-        except Exception as e:
-            self._websocket = None
-            await self.push_error(error_msg=f"Unknown error occurred: {e}", exception=e)
-            await self._call_event_handler("on_connection_error", f"{e}")
+        return url
 
-    async def _disconnect_websocket(self):
-        try:
-            await self.stop_all_metrics()
-            websocket = self._websocket
-            if websocket:
-                logger.debug("Disconnecting from ElevenLabs")
-                # The multi-stream protocol tears down in two steps: we ask
-                # ElevenLabs to close, then it closes. Wait for its close before
-                # forcing ours, so we don't race the closing handshake (which
-                # otherwise ends a notable fraction of sessions in a 1006 close).
-                # The timeout is only a fallback ceiling; the clean close
-                # normally arrives well within it.
-                await websocket.send(json.dumps({"close_socket": True}))
-                try:
-                    await asyncio.wait_for(websocket.wait_closed(), timeout=2.0)
-                except TimeoutError:
-                    logger.debug(
-                        "ElevenLabs did not close the WebSocket within 2.0s; closing from our side"
-                    )
-                await websocket.close()
-                logger.debug("Disconnected from ElevenLabs")
-        except websockets.ConnectionClosed as e:
-            # The server closed the connection first — normal during teardown, or a race
-            # on the closing handshake. The connection is gone either way; this is not a
-            # pipeline error, so don't push an ErrorFrame (which would e.g. trigger a
-            # spurious ServiceSwitcherStrategyFailover switch during shutdown).
-            logger.debug(f"{self} websocket already closed during disconnect: {e}")
-        except Exception as e:
-            await self.push_error(error_msg=f"Unknown error occurred: {e}", exception=e)
-        finally:
-            await self.remove_active_audio_context()
-            self._websocket = None
-            self._context_init_sent.clear()
-            await self._call_event_handler("on_disconnected")
-
-    def _get_websocket(self):
-        if self._websocket:
-            return self._websocket
-        raise Exception("Websocket not connected")
+    def _clear_connection_state(self):
+        self._context_init_sent.clear()
 
     async def _close_context(self, context_id: str):
         # ElevenLabs requires that Pipecat explicitly closes contexts to free
@@ -876,22 +507,9 @@ class ElevenLabsTTSService(WebsocketTTSService):
                 await self.push_error(error_msg=f"Unknown error occurred: {e}", exception=e)
 
     def _reset_alignment_state(self, context_id: str):
-        self._cumulative_time = 0.0
-        self._partial_word = ""
-        self._partial_word_start_time = 0.0
+        super()._reset_alignment_state(context_id)
         self._alignment_started_context_ids.discard(context_id)
         self._context_init_sent.discard(context_id)
-
-    async def on_audio_context_interrupted(self, context_id: str):
-        """Close the ElevenLabs context when the bot is interrupted."""
-        await self._close_context(context_id)
-        self._reset_alignment_state(context_id)
-        await super().on_audio_context_interrupted(context_id)
-
-    async def on_audio_context_completed(self, context_id: str):
-        """Reset alignment state after all audio for the context has played."""
-        self._reset_alignment_state(context_id)
-        await super().on_audio_context_completed(context_id)
 
     async def on_turn_context_completed(self):
         """Close the server-side context at end of turn.
@@ -977,17 +595,6 @@ class ElevenLabsTTSService(WebsocketTTSService):
                             "_receive_messages: using fallback timing method - consider investigating alignment data structure"
                         )
 
-    async def _keepalive_task_handler(self):
-        """Send periodic keepalive messages to maintain WebSocket connection."""
-        KEEPALIVE_SLEEP = 10
-        while True:
-            await asyncio.sleep(KEEPALIVE_SLEEP)
-            try:
-                await self._send_keepalive()
-            except websockets.ConnectionClosed as e:
-                logger.warning(f"{self} keepalive error: {e}")
-                break
-
     async def _send_keepalive(self):
         """Send a single keepalive message to keep the WebSocket connection alive.
 
@@ -1012,65 +619,25 @@ class ElevenLabsTTSService(WebsocketTTSService):
             keepalive_message = {"text": ""}
         await self._websocket.send(json.dumps(keepalive_message))
 
+    async def _send_context_init(self, context_id: str):
+        """Open a context, carrying voice settings and pronunciation dictionaries."""
+        # Mark the context-init as sent so the keepalive may now target this
+        # context_id.
+        self._context_init_sent.add(context_id)
+        msg: dict[str, Any] = {"text": " ", "context_id": context_id}
+        if self._voice_settings:
+            msg["voice_settings"] = self._voice_settings
+        if self._pronunciation_dictionary_locators:
+            msg["pronunciation_dictionary_locators"] = [
+                locator.model_dump() for locator in self._pronunciation_dictionary_locators
+            ]
+        await self._get_websocket().send(json.dumps(msg))
+
     async def _send_text(self, text: str, context_id: str):
         """Send text to the WebSocket for synthesis."""
         if self._websocket and context_id:
             msg = {"text": text, "context_id": context_id}
             await self._websocket.send(json.dumps(msg))
-
-    @traced_tts
-    async def run_tts(self, text: str, context_id: str) -> AsyncGenerator[Frame | None, None]:
-        """Generate speech from text using ElevenLabs' streaming WebSocket API.
-
-        Args:
-            text: The text to synthesize into speech.
-            context_id: The context ID for tracking audio frames.
-
-        Yields:
-            Frame: Audio frames containing the synthesized speech.
-        """
-        try:
-            if not self._websocket or self._websocket.state is State.CLOSED:
-                await self._connect()
-
-            if self._websocket is None:
-                logger.warning(f"{self}: websocket unavailable after reconnect, skipping TTS")
-                yield ErrorFrame(error="websocket unavailable")
-                return
-
-            try:
-                if not self.audio_context_available(context_id):
-                    await self.create_audio_context(context_id)
-                    await self.start_ttfb_metrics()
-                    yield TTSStartedFrame(context_id=context_id)
-                    self._cumulative_time = 0
-                    self._partial_word = ""
-                    self._partial_word_start_time = 0.0
-
-                    # Initialize context with voice settings and pronunciation dictionaries
-                    msg: dict[str, Any] = {"text": " ", "context_id": context_id}
-                    if self._voice_settings:
-                        msg["voice_settings"] = self._voice_settings
-                    if self._pronunciation_dictionary_locators:
-                        msg["pronunciation_dictionary_locators"] = [
-                            locator.model_dump()
-                            for locator in self._pronunciation_dictionary_locators
-                        ]
-                    # Mark the context-init as sent so the keepalive may now
-                    # target this context_id.
-                    self._context_init_sent.add(context_id)
-                    await self._websocket.send(json.dumps(msg))
-                    logger.trace(f"Created new context {context_id}")
-
-                await self._send_text(text, context_id)
-                await self.start_tts_usage_metrics(text)
-            except Exception as e:
-                yield TTSStoppedFrame(context_id=context_id)
-                yield ErrorFrame(error=f"Unknown error occurred: {e}")
-                return
-            yield None
-        except Exception as e:
-            yield ErrorFrame(error=f"Unknown error occurred: {e}")
 
 
 class ElevenLabsHttpTTSService(TTSService):
